requirement.txt
streamlit
langchain
langchain_community
python-dotenv
transformers
sentence-transformers
pyMuPDF

